Refactor out some ratings prompt logic from SuccessPanel to RatingsHelper
